+++ Space Invaders/Space Invaders/barrierblock.cpp
@@ -21,7 +21,7 @@
 
 
 
-CBarrierBlock::CBarrierBlock(const int _iX, const int _iY)
+CBarrierBlock::CBarrierBlock(const float _iX, const float _iY)
 {
 	m_fX = _iX;
 	m_fY = _iY;
